+++ neoforge/src/main/java/earth/terrarium/botarium/neoforge/data/NeoDataManager.java
@@ -31,11 +31,6 @@ public class NeoDataManager<T> implements DataManager<T> {
     }
 
     @Override
-    public T getDataOrInit(Object dataHolder) {
-        return getData(dataHolder);
-    }
-
-    @Override
     public T getDataOrInit(Object dataHolder, T data) {
         if (!hasData(dataHolder)) {
             setData(dataHolder, data);