+++ changelog.md
@@ -1,3 +1,3 @@
 # 3.2.3
 
-- Implement Attachments on ItemStacks
+- getData now inits on fabric